requirements.txt
streamlit==1.29.0
openai==1.6.1
loguru==0.7.2
pyperclip==1.8.2
python-dotenv==1.0.0